Adds constraint to ensure unique messages
This adds an 'event_hash' field to the database of type BYTEA - it
should be populated with an ID that is either an id form the
source (e.g., github's notification id) or a suitable hash to ensure
uniqueness.  event_hash + source_identifier  ensure that trackers do not
insert the same message twice.

This also adds a persistent db connection objct in configObject -
otherwise if get_pg_db returns a new connection each time then the
database object in Event()  isn't the same one as in notification() and
db.rollback() won't work correctly on IntegrityErrors etc.