Move authentication into its own section in config
Individual tracker configs are passed as arguments to celery and in the
case of celerybeat they are persisted to either a shelve file (by
default) or to a database (not yet implemented).  This means any
authentication information in the tracker config is stored in plaintext
or in the database.  This commit moves authentication information into
its own section in the config so that we can prevent it from leaking
into any serialized objects.

This also adds some basic validation for task configs and clears up some
of the naming in the github notification tracker.